requirements.txt
Flask==2.3.2
google-cloud-speech==2.21.0
google-cloud-texttospeech==2.14.1